Chore : 계정 설정페이지 파일경로 수정, 헤더 자유게시판 경로 재수정
