+++ src/components/layout/gnb/SideMenu.tsx
@@ -4,7 +4,6 @@
 import React, { forwardRef } from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
-import closeIcon from '@/../public/icons/close.svg';
 
 interface Team {
   id: number;
@@ -29,10 +28,10 @@ const SideMenu = forwardRef<HTMLDivElement, SideMenuProps>(({ teams, isOpen, onC
       {/* 사이드 메뉴 */}
       <div
         ref={ref}
-        className="bg-bg200 fixed top-0 left-0 z-300 flex h-screen w-[204px] flex-col gap-6 overflow-scroll p-4 pb-10 shadow-lg"
+        className="bg-bg200 fixed top-0 left-0 z-300 flex h-screen w-51 flex-col gap-6 overflow-scroll p-4 pb-10 shadow-lg"
       >
-        <button onClick={onClose} className="cursor-pointer self-end" title="닫기">
-          <Image src={closeIcon} alt="닫기" width={24} height={24} />
+        <button type="button" onClick={onClose} className="cursor-pointer self-end" title="닫기">
+          <Image src="/icons/close.svg" alt="닫기" width={24} height={24} />
         </button>
 
         {/* 팀 목록 */}
